[Feature] ORG 정보 수정 0.2.1

+++ src/main/java/metaint/replanet/rest/org/repository/OrgMemberRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrgMemberRepository extends JpaRepository<Member, Integer> {
+public interface OrgMemberRepository extends JpaRepository<Member, Long> {
     List<Member> findAllByMemberRole(MemberRole memberRole);
 }
